refactor: separate group selector
