+++ hwproj.front/src/components/Homeworks/CourseHomeworkExperimental.tsx
@@ -11,7 +11,6 @@
     TextField,
     Tooltip,
     Typography,
-    Autocomplete
 } from "@mui/material";
 import {MarkdownEditor, MarkdownPreview} from "components/Common/MarkdownEditor";
 import FilesPreviewList from "components/Files/FilesPreviewList";
@@ -19,7 +18,7 @@ import {IFileInfo} from "components/Files/IFileInfo";
 import {FC, useEffect, useState} from "react"
 import Utils from "services/Utils";
 import {
-    HomeworkViewModel, ActionOptions, HomeworkTaskViewModel, PostTaskViewModel, Group
+    HomeworkViewModel, ActionOptions, HomeworkTaskViewModel, PostTaskViewModel
 } from "@/api";
 import ApiSingleton from "../../api/ApiSingleton";
 import Tags from "../Common/Tags";
@@ -38,6 +37,7 @@ import Lodash from "lodash";
 import {CourseUnitType} from "../Files/CourseUnitType"
 import ProcessFilesUtils from "../Utils/ProcessFilesUtils";
 import {FilesHandler} from "@/components/Files/FilesHandler";
+import GroupSelector from "../Common/GroupSelector";
 
 export interface HomeworkAndFilesInfo {
     homework: HomeworkViewModel & { isModified?: boolean },
@@ -116,8 +116,6 @@ const CourseHomeworkEditor: FC<{
     const [tags, setTags] = useState<string[]>(loadedHomework.tags!)
     const [description, setDescription] = useState<string>(loadedHomework.description!)
     const [selectedGroupId, setSelectedGroupId] = useState(loadedHomework.groupId)
-    const [groups, setGroups] = useState<Group[]>([])
-    const [groupsLoading, setGroupsLoading] = useState(false)
 
     const [hasErrors, setHasErrors] = useState<boolean>(false)
 
@@ -127,21 +125,6 @@ const CourseHomeworkEditor: FC<{
 
     const [deadlineSuggestion, setDeadlineSuggestion] = useState<Date | undefined>(undefined)
     const [tagSuggestion, setTagSuggestion] = useState<string | undefined>(undefined)
-
-    const loadGroups = async () => {
-        setGroupsLoading(true)
-        try {
-            const courseGroups = await ApiSingleton.courseGroupsApi.courseGroupsGetAllCourseGroupsWithNames(courseId)
-            setGroups(courseGroups)
-        } catch (error) {
-            console.error('Failed to load groups:', error)
-        } finally {
-            setGroupsLoading(false)
-        }
-    }
-    useEffect(() => {
-        loadGroups()
-    }, [courseId])
 
     useEffect(() => {
         if (!isNewHomework || !metadata.publicationDate) return
@@ -308,37 +291,12 @@ const CourseHomeworkEditor: FC<{
                 </Grid>
             </Grid>
             <Grid container>
-                <Grid item xs={12} style={{marginBottom: "15px", marginTop: 1}}>
-                    {!isNewHomework && isPublished ? (
-                        <TextField
-                            label="Группа"
-                            value={groups.find(g => g.id === loadedHomework.groupId)?.name || "Все студенты"}
-                            variant="outlined"
-                            fullWidth
-                            disabled
-                        />
-                    ) : (
-                        <Autocomplete
-                            options={[{ id: undefined, name: "Все студенты" }, ...groups]}
-                            getOptionLabel={(option) => option.name || ""}
-                            value={selectedGroupId !== undefined
-                                ? groups.find(g => g.id === selectedGroupId) || null
-                                : { id: undefined, name: "Все студенты" }}
-                            onChange={(event, newValue) => {
-                                setSelectedGroupId(newValue?.id)
-                            }}
-                            loading={groupsLoading}
-                            renderInput={(params) => (
-                                <TextField
-                                    {...params}
-                                    label="Группа (не обязательно)"
-                                    placeholder="Выберите группу"
-                                    variant="outlined"
-                                />
-                            )}
-                        />
-                    )}
-                </Grid>
+                <GroupSelector
+                    courseId={courseId}
+                    onGroupIdChange={(groupId?: number) => setSelectedGroupId(groupId)}
+                    selectedGroupId={selectedGroupId}
+                    disabled={!isNewHomework}
+                />
                 {tags.includes(TestTag) &&
                     <Grid item>
                         <Alert severity="info" variant={"outlined"}>
